Hot fix

+++ src/Project/repositories/Account.js
@@ -33,7 +33,7 @@ export default class Account {
     async edit(id, data) {
         const promise = []
         const account = await this.getAccountByID(id)
-        if (account.photo) {
+        if (data.photo && account.photo) {
             promise.push(utils.removeFile(utils.getPath('./uploadedFiles', account.photo)))
         }
         promise.push(this.AccountModel.where('id', id).update(data))
